commit on Week-0

+++ main-3-5.cpp
@@ -5,12 +5,12 @@ using namespace std;
 extern double sumeven(double[], int);
 
 int main(int argc, char **argv){
-    int arr[argc];
+    double arr[argc];
 
     for(int i =0; i < argc; i++){
         arr[i] = *(argv[i]);
     }
-    
+
     sumeven(arr, argc);
 
     return 0;
